update transect.py
1. rewrite script using _readfile.read() and argparse()
2. add —line-file option to read start/end point from GMT point file,
useful for multiple files sharing the same profile file
3. move manual select point into function
manual_select_start_end_point()
4. add transect_yx() and transect_lalo() to get transect easily,
supporting different resolutions
5. plot data profile and DEM in the same figure.

Since i have not finish the rewrite of GPS input file and profile
buffer, I renamed old transect.py to transect_legacy.py for future
reference.
